clean up the `nkBracketExpr` handling for calls
Since `nkBracketExpr` nodes are no longer rewritten into calls of the
`[]` procedure, attempting to reconstruct the original is no longer
necessary.

This is partial revert of the following commits:
- #8ef66b973d86a75c8dfa4c6761d322d94c54efad
- #e01e0a4b4056ac4a42797bab9f10822ae8996607

In addition, `mArrGet` no longer being used for all forms of bracket
expressions allows for removing the magic's special casing in
`sigmatch` and `semmagic`, making it act more like a normal symbol.