macros: fix `typeKind` for `NimNode` without type
Summary
=======

Fix `typeKind` used on `NimNode`s without a type yielding an unstable
and potentially invalid value. For a `NimNode` without a type,
`typeKind` is now guaranteed to return `ntyNone`.

Details
=======

The opcode implementation didn't set the result register (`ra`) value
to `ntyNone` when the node had no associated type. If the result
register is a re-used `rkInt` register (that is, one for which no
register transition is required), this resulted in `typeKind` returning
an arbitrary value.

If the `NimNode` operand has no type associated, the result value is
now explicitly set to `ord(tyNone)`.